Theoretically async_file_handle now supports the Coroutines TS. I have no way of testing it until MSVC stops ICEing on AFIO (VS2017 15.5 they say)

+++ include/afio/v2.0/detail/impl/windows/utils.ipp
@@ -188,7 +188,7 @@ namespace utils
     // Flush modified data so dropping the cache drops everything
     OUTCOME_TRYV(flush_modified_data());
     // Drop filesystem cache
-    if(!SetSystemFileCacheSize(-1, -1, 0))
+    if(!SetSystemFileCacheSize((SIZE_T) -1, (SIZE_T) -1, 0))
     {
       return {GetLastError(), std::system_category()};
     }
